test_DiffIterate.cc
